Create workspace with 4 members
Add eval subcommand
