+++ Cargo.toml
@@ -1,16 +1,7 @@
-[package]
-name = "spiderfire"
-version = "0.1.1"
-authors = ["Redfire <redfire75369@hotmail.com>"]
-edition = "2018"
-
-
-[dependencies]
-chrono = "0.4.19"
-clap = "2.33.3"
-libc = "0.2.92"
-mozjs = { package = "mozjs", git = "https://github.com/servo/rust-mozjs", features = ["debugmozjs"] }
-once_cell = "1.7.2"
-# rb_tree = { package = "rb_tree", git = "https://github.com/Redfire75369/rb_tree" }
-# rb_tree = "0.3.5"
-termcolor = "1.1.2"
+[workspace]
+members = [
+	"cli",
+	"ion",
+	"modules",
+	"runtime"
+]
