+++ build.bat
@@ -3,6 +3,7 @@
 set MOZTOOLS_PATH=C:\mozilla-build\msys\bin;C:\mozilla-build\bin
 set AUTOCONF=C:\mozilla-build\msys\local\bin\autoconf-2.13
 set NATIVE_WIN32_PYTHON=C:\mozilla-build\python\python2.7.exe
+set PYTHON3=C:\mozilla-build\python3\python3.exe
 
 set LIBCLANG_PATH=C:\Program Files\LLVM\lib
 set CC=clang-cl.exe
