+++ cli/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cli"
+version = "0.1.0"
+authors = ["Redfire <redfire75369@hotmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = "2.33.3"
+ion = { path = "../ion" }
+modules = { path = "../modules" }
+mozjs = { package = "mozjs", git = "https://github.com/servo/rust-mozjs", features = ["debugmozjs"] }
+once_cell = "1.7.2"
+runtime = { path = "../runtime" }
