+++ cli/src/commands/eval.rs
@@ -1,0 +1,27 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+use ::std::ptr;
+
+use mozjs::jsapi::*;
+use mozjs::rust::{JSEngine, RealmOptions, Runtime, SIMPLE_GLOBAL_CLASS};
+
+use crate::evaluate::eval_inline;
+
+pub fn eval_source(source: &str) {
+	let engine = JSEngine::init().expect("JS Engine Initialisation Failed");
+	let rt = Runtime::new(engine.handle());
+
+	assert!(!rt.cx().is_null(), "JSContext Creation Failed");
+
+	let h_options = OnNewGlobalHookOption::FireOnNewGlobalHook;
+	let c_options = RealmOptions::default();
+
+	let global = unsafe { JS_NewGlobalObject(rt.cx(), &SIMPLE_GLOBAL_CLASS, ptr::null_mut(), h_options, &*c_options) };
+
+	let _ac = JSAutoRealm::new(rt.cx(), global);
+
+	eval_inline(&rt, global, source);
+}
