+++ cli/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod eval;
+pub mod repl;
+pub mod run;
