+++ ion/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "ion"
+version = "0.1.0"
+authors = ["Redfire <redfire75369@hotmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+libc = "0.2.92"
+mozjs = { package = "mozjs", git = "https://github.com/servo/rust-mozjs", features = ["debugmozjs"] }
+termcolor = "1.1.2"
