+++ runtime/src/lib.rs
@@ -1,0 +1,6 @@
+#[macro_use]
+extern crate mozjs;
+
+pub mod globals;
+pub mod init;
+pub mod modules;
