--- src/runtime/jsapi_utils/mod.rs
@@ -1,12 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-pub mod array;
-pub mod eval;
-mod exception;
-pub mod functions;
-pub mod print;
-pub mod string;
